+++ packages/api/lib/writeRecords/write-granules.js
@@ -2,19 +2,20 @@
 
 const AggregateError = require('aggregate-error');
 const isEmpty = require('lodash/isEmpty');
+const omit = require('lodash/omit');
 const pMap = require('p-map');
 
 const { s3 } = require('@cumulus/aws-client/services');
 const cmrUtils = require('@cumulus/cmrjs/cmr-utils');
 const { buildURL } = require('@cumulus/common/URLUtils');
 const {
-  translateApiFiletoPostgresFile,
-  FilePgModel,
-  GranulePgModel,
-  upsertGranuleWithExecutionJoinRecord,
-  translateApiGranuleToPostgresGranule,
   CollectionPgModel,
   createRejectableTransaction,
+  FilePgModel,
+  GranulePgModel,
+  translateApiFiletoPostgresFile,
+  translateApiGranuleToPostgresGranule,
+  upsertGranuleWithExecutionJoinRecord,
 } = require('@cumulus/db');
 const {
   upsertGranule,
@@ -356,6 +357,74 @@ const writeGranuleToDynamoAndEs = async (params) => {
   }
 };
 
+const _writeGranuleRecords = async ({
+  postgresGranuleRecord,
+  apiGranuleRecord,
+  knex,
+  esClient,
+  granuleModel,
+  executionCumulusId,
+  granulePgModel,
+}) => {
+  let pgGranule;
+  log.info('About to write granule record %j to PostgreSQL', postgresGranuleRecord);
+  log.info('About to write granule record %j to DynamoDB', apiGranuleRecord);
+
+  await createRejectableTransaction(knex, async (trx) => {
+    pgGranule = await _writePostgresGranuleViaTransaction({
+      granuleRecord: postgresGranuleRecord,
+      executionCumulusId,
+      trx,
+      granulePgModel,
+    });
+    await writeGranuleToDynamoAndEs({
+      apiGranuleRecord,
+      esClient,
+      granuleModel,
+    });
+  });
+  log.info(
+    `
+    Successfully wrote granule %j to PostgreSQL. Record cumulus_id in PostgreSQL: ${pgGranule.cumulus_id}.
+    `,
+    postgresGranuleRecord
+  );
+  log.info('Successfully wrote granule %j to DynamoDB', apiGranuleRecord);
+  return pgGranule;
+};
+
+const _writePostgresFilesFromApiGranuleFiles = async ({
+  apiGranuleRecord,
+  granuleCumulusId,
+  knex,
+  snsEventType,
+}) => {
+  const { files, granuleId, status, error } = apiGranuleRecord;
+  if (status !== 'running' && status !== 'queued' && files.length > 0) {
+    await _writeGranuleFiles({
+      files,
+      granuleCumulusId: granuleCumulusId,
+      granuleId,
+      workflowError: error,
+      knex,
+      snsEventType,
+      granuleModel: new Granule(),
+    });
+  }
+};
+
+const _publishPostgresGranuleUpdateToSns = async ({
+  snsEventType,
+  pgGranule,
+  knex,
+}) => {
+  const granuletoPublish = await translatePostgresGranuleToApiGranule({
+    granulePgRecord: pgGranule,
+    knexOrTransaction: knex,
+  });
+  await publishGranuleSnsMessageByEventType(granuletoPublish, snsEventType);
+};
+
 /**
  * Write a granule record to DynamoDB and PostgreSQL
  *
@@ -381,86 +450,188 @@ const _writeGranule = async ({
   knex,
   snsEventType,
 }) => {
-  let pgGranule;
+  const pgGranule = await _writeGranuleRecords({
+    postgresGranuleRecord,
+    apiGranuleRecord,
+    knex,
+    esClient,
+    granuleModel,
+    executionCumulusId,
+    granulePgModel,
+  });
 
-  log.info('About to write granule record %j to PostgreSQL', postgresGranuleRecord);
-  log.info('About to write granule record %j to DynamoDB', apiGranuleRecord);
+  await _writePostgresFilesFromApiGranuleFiles({
+    apiGranuleRecord,
+    granuleCumulusId: pgGranule.cumulus_id,
+    knex,
+    snsEventType,
+  });
 
+  await _publishPostgresGranuleUpdateToSns({
+    snsEventType,
+    pgGranule,
+    knex,
+  });
+};
+
+/**
+* Method to facilitate parital granule record updates
+* @summary In cases where a full API record is not passed, but partial/tangential updates to granule
+*          records are called for, updates to files records are not required and pre-write
+*          calculation in methods like write/update GranulesFromApi result in unneded
+*          evaluation/database writes /etc. This method updates the postgres/Dynamo/ES datastore and
+*          publishes the SNS update event without incurring unneded overhead.
+* @param {Object}          params
+* @param {Object}          params.apiGranuleRecord - Api Granule object to write to the database
+* @param {number}          params.executionCumulusId - Execution ID the granule was written from
+* @param {Object}          params.esClient - Elasticsearch client
+* @param {Object}          params.granuleModel - Instance of DynamoDB granule model
+* @param {Object}          params.granulePgModel - @cumulus/db compatible granule module instance
+* @param {Knex}            params.knex - Knex object
+* @param {Object}          params.postgresGranuleRecord - PostgreSQL granule record to write
+*                                                         to database
+* @param {string}          params.snsEventType - SNS Event Type
+* @returns {Promise}
+*/
+const writeGranuleRecordAndPublishSns = async ({
+  postgresGranuleRecord,
+  apiGranuleRecord,
+  esClient,
+  executionCumulusId,
+  granuleModel,
+  granulePgModel,
+  knex,
+  snsEventType = 'Update',
+}) => {
+  const pgGranule = await _writeGranuleRecords({
+    postgresGranuleRecord,
+    apiGranuleRecord,
+    knex,
+    esClient,
+    granuleModel,
+    executionCumulusId,
+    granulePgModel,
+  });
+  await _publishPostgresGranuleUpdateToSns({
+    snsEventType,
+    pgGranule,
+    knex,
+  });
+};
+
+/**
+ * Update granule record status in DynamoDB, PostgreSQL, Elasticsearch.
+ * Publish SNS event for updated granule.
+ *
+ * @param {Object}  params
+ * @param {Object}  params.apiGranule            - API Granule object to write to
+ *                                                 the database
+ * @param {Object}  params.postgresGranule       - PostgreSQL granule
+ * @param {Object}  params.apiFieldUpdates       - API fields to update
+ * @param {Object}  params.pgFieldUpdates        - PostgreSQL fields to update
+ * @param {Object}  params.apiFieldsToDelete     - API fields to delete
+ * @param {Object}  params.granuleModel          - Instance of DynamoDB granule model
+ * @param {Object}  params.granulePgModel        - @cumulus/db compatible granule module instance
+ * @param {Knex}    params.knex                  - Knex object
+ * @param {string}  params.snsEventType          - SNS Event Type, defaults to 'Update'
+ * @param {Object}  params.esClient              - Elasticsearch client
+ * returns {Promise}
+ */
+const _updateGranule = async ({
+  apiGranule,
+  postgresGranule,
+  apiFieldUpdates,
+  pgFieldUpdates,
+  apiFieldsToDelete,
+  granuleModel,
+  granulePgModel,
+  knex,
+  snsEventType = 'Update',
+  esClient,
+}) => {
+  const granuleId = apiGranule.granuleId;
+  const esGranule = omit(apiGranule, apiFieldsToDelete);
+
+  let updatedPgGranule;
   await createRejectableTransaction(knex, async (trx) => {
-    pgGranule = await _writePostgresGranuleViaTransaction({
-      granuleRecord: postgresGranuleRecord,
-      executionCumulusId,
+    [updatedPgGranule] = await granulePgModel.update(
       trx,
-      granulePgModel,
-    });
-    await writeGranuleToDynamoAndEs({
-      apiGranuleRecord,
-      esClient,
-      granuleModel,
-    });
+      { cumulus_id: postgresGranule.cumulus_id },
+      pgFieldUpdates,
+      ['*']
+    );
+    log.info(`Successfully wrote granule ${granuleId} to PostgreSQL`);
+    try {
+      // delete execution field
+      await granuleModel.update({ granuleId }, apiFieldUpdates, apiFieldsToDelete);
+      log.info(`Successfully wrote granule ${granuleId} to DynamoDB`);
+      await upsertGranule({
+        esClient,
+        updates: {
+          ...esGranule,
+          ...apiFieldUpdates,
+        },
+        index: process.env.ES_INDEX,
+      });
+      log.info(`Successfully wrote granule ${granuleId} to Elasticsearch`);
+    } catch (writeError) {
+      log.info(`Writes to DynamoDB/Elasticsearch failed, rolling back all writes for granule ${granuleId}`);
+      // On error, recreate the DynamoDB record to revert it back to original
+      // status to ensure that all systems stay in sync
+      await granuleModel.create(apiGranule);
+      throw writeError;
+    }
   });
 
   log.info(
     `
-    Successfully wrote granule %j to PostgreSQL. Record cumulus_id in PostgreSQL: ${pgGranule.cumulus_id}.
+    Successfully wrote granule %j to PostgreSQL. Record cumulus_id in PostgreSQL: ${updatedPgGranule.cumulus_id}.
     `,
-    postgresGranuleRecord
+    postgresGranule
   );
-  log.info('Successfully wrote granule %j to DynamoDB', apiGranuleRecord);
-
-  const { files, granuleId, status, error } = apiGranuleRecord;
-
-  if (status !== 'running' && status !== 'queued' && files.length > 0) {
-    await _writeGranuleFiles({
-      files,
-      granuleCumulusId: pgGranule.cumulus_id,
-      granuleId,
-      workflowError: error,
-      knex,
-      snsEventType,
-      granuleModel,
-    });
-  }
+  log.info('Successfully wrote granule %j to DynamoDB', apiGranule);
 
   const granuletoPublish = await translatePostgresGranuleToApiGranule({
-    granulePgRecord: pgGranule,
+    granulePgRecord: updatedPgGranule,
     knexOrTransaction: knex,
   });
   await publishGranuleSnsMessageByEventType(granuletoPublish, snsEventType);
+  log.info('Successfully wrote granule %j to SNS topic', granuletoPublish);
 };
 
 /**
  * Thin wrapper to _writeGranule used by endpoints/granule to create a granule
  * directly.
  *
- * @param {Object} params
- * @param {string} params.granuleId - granule's id
- * @param {string} params.collectionId - granule's collection id
- * @param {GranuleStatus} params.status - ['running','failed','completed', 'queued']
- * @param {string} [params.execution] - Execution URL to associate with this granule
+ * @param {Object} granule -- API Granule object
+ * @param {string} granule.granuleId - granule's id
+ * @param {string} granule.collectionId - granule's collection id
+ * @param {GranuleStatus} granule.status - ['running','failed','completed', 'queued']
+ * @param {string} [granule.execution] - Execution URL to associate with this granule
  *                               must already exist in database.
- * @param {string} [params.cmrLink] - url to CMR information for this granule.
- * @param {boolean} [params.published] - published to cmr
- * @param {string} [params.pdrName] - pdr name
- * @param {string} [params.provider] - provider
- * @param {Object} [params.error = {}] - workflow errors
- * @param {string} [params.createdAt = new Date().valueOf()] - time value
- * @param {string} [params.timestamp] - timestamp
- * @param {string} [params.updatedAt = new Date().valueOf()] - time value
- * @param {number} [params.duration] - seconds
- * @param {string} [params.productVolume] - sum of the files sizes in bytes
- * @param {integer} [params.timeToPreprocess] -  seconds
- * @param {integer} [params.timeToArchive] - seconds
- * @param {Array<ApiFile>} params.files - files associated with the granule.
- * @param {string} [params.beginningDateTime] - CMR Echo10: Temporal.RangeDateTime.BeginningDateTime
- * @param {string} [params.endingDateTime] - CMR Echo10: Temporal.RangeDateTime.EndingDateTime
- * @param {string} [params.productionDateTime] - CMR Echo10: DataGranule.ProductionDateTime
- * @param {string} [params.lastUpdateDateTime] - CMR Echo10: LastUpdate || InsertTime
- * @param {string} [params.processingStartDateTime] - execution startDate
- * @param {string} [params.processingEndDateTime] - execution StopDate
- * @param {Object} [params.queryFields] - query fields
- * @param {Object} [params.granuleModel] - only for testing.
- * @param {Object} [params.granulePgModel] - only for testing.
+ * @param {string} [granule.cmrLink] - url to CMR information for this granule.
+ * @param {boolean} [granule.published] - published to cmr
+ * @param {string} [granule.pdrName] - pdr name
+ * @param {string} [granule.provider] - provider
+ * @param {Object} [granule.error = {}] - workflow errors
+ * @param {string} [granule.createdAt = new Date().valueOf()] - time value
+ * @param {string} [granule.timestamp] - timestamp
+ * @param {string} [granule.updatedAt = new Date().valueOf()] - time value
+ * @param {number} [granule.duration] - seconds
+ * @param {string} [granule.productVolume] - sum of the files sizes in bytes
+ * @param {integer} [granule.timeToPreprocess] -  seconds
+ * @param {integer} [granule.timeToArchive] - seconds
+ * @param {Array<ApiFile>} granule.files - files associated with the granule.
+ * @param {string} [granule.beginningDateTime] - CMR Echo10:
+ *                                               Temporal.RangeDateTime.BeginningDateTime
+ * @param {string} [granule.endingDateTime] - CMR Echo10: Temporal.RangeDateTime.EndingDateTime
+ * @param {string} [granule.productionDateTime] - CMR Echo10: DataGranule.ProductionDateTime
+ * @param {string} [granule.lastUpdateDateTime] - CMR Echo10: LastUpdate || InsertTime
+ * @param {string} [granule.processingStartDateTime] - execution startDate
+ * @param {string} [granule.processingEndDateTime] - execution StopDate
+ * @param {Object} [granule.queryFields] - query fields
+ * @param {Object} [granule.granuleModel] - only for testing.
+ * @param {Object} [granule.granulePgModel] - only for testing.
  * @param {Knex} knex - knex Client
  * @param {Object} esClient - Elasticsearch client
  * @param {string} snsEventType - SNS Event Type
@@ -689,15 +860,16 @@ const writeGranulesFromMessage = async ({
  * @returns {Promise}
  * @throws {Error}
  */
-async function updateGranuleStatusToQueued({
-  granule,
-  knex,
-  collectionPgModel = new CollectionPgModel(),
-  granuleModel = new Granule(),
-  granulePgModel = new GranulePgModel(),
-}) {
+const updateGranuleStatusToQueued = async (params) => {
+  const {
+    granule,
+    knex,
+    collectionPgModel = new CollectionPgModel(),
+    granuleModel = new Granule(),
+    granulePgModel = new GranulePgModel(),
+    esClient = await Search.es(),
+  } = params;
   const { granuleId, collectionId } = granule;
-  const status = 'queued';
   log.info(`updateGranuleStatusToQueued granuleId: ${granuleId}, collectionId: ${collectionId}`);
 
   try {
@@ -705,7 +877,7 @@ async function updateGranuleStatusToQueued({
       knex,
       deconstructCollectionId(collectionId)
     );
-    const granuleCumulusId = await granulePgModel.getRecordCumulusId(
+    const pgGranule = await granulePgModel.get(
       knex,
       {
         granule_id: granuleId,
@@ -713,26 +885,34 @@ async function updateGranuleStatusToQueued({
       }
     );
 
-    await createRejectableTransaction(knex, async (trx) => {
-      await granulePgModel.update(trx, { cumulus_id: granuleCumulusId }, { status });
-      // delete the execution field as well
-      await granuleModel.update({ granuleId }, { status }, ['execution']);
+    await _updateGranule({
+      apiGranule: granule,
+      postgresGranule: pgGranule,
+      apiFieldUpdates: { status: 'queued' },
+      pgFieldUpdates: { status: 'queued' },
+      apiFieldsToDelete: ['execution'],
+      granuleModel,
+      granulePgModel,
+      knex,
+      snsEventType: 'Update',
+      esClient,
     });
 
-    log.debug(`Updated granule status to queued, Dynamo granuleId: ${granule.granuleId}, PostgreSQL cumulus_id: ${granuleCumulusId}`);
+    log.debug(`Updated granule status to queued, Dynamo granuleId: ${granule.granuleId}, PostgreSQL cumulus_id: ${pgGranule.cumulus_id}`);
   } catch (thrownError) {
     log.error(`Failed to update granule status to queued, granuleId: ${granule.granuleId}, collectionId: ${collectionId}`, thrownError);
     throw thrownError;
   }
-}
+};
 
 module.exports = {
   _writeGranule,
+  createGranuleFromApi,
   generateFilePgRecord,
-  updateGranuleStatusToQueued,
   getGranuleFromQueryResultOrLookup,
+  updateGranuleFromApi,
+  updateGranuleStatusToQueued,
   writeGranuleFromApi,
   writeGranulesFromMessage,
-  createGranuleFromApi,
-  updateGranuleFromApi,
+  writeGranuleRecordAndPublishSns,
 };