+++ packages/api/tests/endpoints/granules/test-bulk-reingest.js
@@ -98,6 +98,7 @@ test.serial('POST /granules/bulkReingest starts an async-operation with the corr
     payload: body,
     type: 'BULK_GRANULE_REINGEST',
     envVars: {
+      ES_HOST: process.env.ES_HOST,
       GranulesTable: process.env.GranulesTable,
       granule_sns_topic_arn: process.env.granule_sns_topic_arn,
       system_bucket: process.env.system_bucket,
@@ -147,6 +148,7 @@ test.serial('POST /granules/bulkReingest starts an async-operation with the corr
     payload: body,
     type: 'BULK_GRANULE_REINGEST',
     envVars: {
+      ES_HOST: process.env.ES_HOST,
       GranulesTable: process.env.GranulesTable,
       granule_sns_topic_arn: process.env.granule_sns_topic_arn,
       system_bucket: process.env.system_bucket,