+++ CHANGELOG.md
@@ -49,6 +49,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - **CUMULUS-2311** - RDS Migration Epic Phase 2
+  - **CUMULUS-2769**
+    - Update collection PUT endpoint to require existance of postgresql record
+      and to ignore lack of dynamoDbRecord on update
+  - **CUMULUS-2767**
+    - Update provider PUT endpoint to require existence of PostgreSQL record
+      and to ignore lack of DynamoDB record on update
   - **CUMULUS-2759**
     - Updates collection/provider/rules/granules creation (post) endpoints to
       primarily check for existence/collision in PostgreSQL database instead of DynamoDB
@@ -128,10 +134,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Updated CreateReconciliationReport lambda to search CMR collections with CMRSearchConceptQueue.
 - **CUMULUS-2638**
   - Adds documentation to clarify bucket config object use.
+- **CUMULUS-2642**
+   - Reduces the reconcilation report's default maxResponseSize that returns
+     the full report rather than an s3 signed url. Reports very close to the
+     previous limits were failing to download, so the limit has been lowered to
+     ensure all files are handled properly.
+- **CUMULUS-2684**
+  - Added optional collection level parameter `s3MultipartChunksizeMb` to collection's `meta` field
+  - Updated `move-granules` task to take in an optional config parameter s3MultipartChunksizeMb
 
 ### Removed
 
 - **CUMULUS-2311** - RDS Migration Epic Phase 2
+  - **CUMULUS-2770**
+    - Removed `waitForModelStatus` from `example/spec/helpers/apiUtils` integration test helpers
   - **CUMULUS-2510**
     - Removed `stream_enabled` and `stream_view_type` from `executions_table` TF
       definition.
@@ -166,6 +182,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Updated SyncGranules to provide `createdAt` on output based on `workflowStartTime` if provided,
   falling back to `Date.now()` if not provided.
   - Updated `task_config` of SyncGranule in example workflows
+- **CUMULUS-2744**
+  - GET executions/status returns associated granules for executions retrieved from the Step Function API
 - **CUMULUS-2751**
   - Upgraded all Cumulus (node.js) workflow tasks to use
     `@cumulus/cumulus-message-adapter-js` version `2.0.3`, which includes an
@@ -337,6 +355,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
       - `rules`
   - **CUMULUS-2294**
     - Updated architecture and deployment documentation to reference RDS
+  - **CUMULUS-2642**
+    - Inventory and Granule Not Found Reconciliation Reports now compare
+      Databse against S3 in on direction only, from Database to S3
+      Objects. This means that only files in the database are compared against
+      objects found on S3 and the filesInCumulus.onlyInS3 report key will
+      always be empty. This significantly decreases the report output size and
+      aligns with a users expectations.
+    - Updates getFilesAndGranuleInfoQuery to take additional optional
+      parameters `collectionIds`, `granuleIds`, and `providers` to allow
+      targeting/filtering of the results.
+
   - **CUMULUS-2694**
     - Updated database write logic in `sfEventSqsToDbRccords` to log message if Cumulus
     workflow message is from pre-RDS deployment but still attempt parallel writing to DynamoDB
@@ -346,6 +375,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     - Updated POST `/executions` endpoint to publish SNS message of created record to executions SNS topic
   - **CUMULUS-2661**
     - Updated PUT `/executions/<arn>` endpoint to publish SNS message of updated record to executions SNS topic
+  - **CUMULUS-2765**
+    - Updated `updateGranuleStatusToQueued` in `write-granules` to write to
+      Elasticsearch and publish SNS message to granules topic.
+  - **CUMULUS-2774**
+    - Updated `constructGranuleSnsMessage` and `constructCollectionSnsMessage`
+      to throw error if `eventType` is invalid or undefined.
 
 ## [v9.9.0] 2021-11-03
 
@@ -372,7 +407,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Added Elasticsearch client scroll setting to the CreateReconciliationReport lambda function.
   - Added `elasticsearch_client_config` tfvars to the archive and cumulus terraform modules.
 - **CUMULUS-2683**
-  - Added `default_s3_multipart_chunksize_mb` setting to the `MoveGranules` lambda function.
+  - Added `default_s3_multipart_chunksize_mb` setting to the `move-granules` lambda function.
   - Added `default_s3_multipart_chunksize_mb` tfvars to the cumulus and ingest terraform modules.
   - Added optional parameter `chunkSize` to `@cumulus/aws-client/S3.moveObject` and
     `@cumulus/aws-client/S3.multipartCopyObject` to set the chunk size of the S3 multipart uploads.